Move template chooser into toolbar
